refactor so that range entry points are in designscript builtin fix designscript builtin resource generation fix tests use constants for function names move enum to designscript builtin
