+++ src/Libraries/DesignScriptBuiltin/Properties/AssemblyInfo.cs
@@ -1,4 +1,5 @@
 ﻿using System.Reflection;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following
@@ -9,3 +10,5 @@ using System.Runtime.InteropServices;
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
 [assembly: Guid("c0d6dee5-5532-4345-9c66-4c00d7fdb8be")]
+
+[assembly:InternalsVisibleTo("ProtoCore")]
